requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Script==2.0.6
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.19.5
pandas==0.25.3
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
pyupbit==0.2.8
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.57.0
websockets==8.0.2
Werkzeug==1.0.1
yarl==1.6.3
